backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
langchain>=0.1.16
langchain-google-genai>=1.0.1
langchain-groq
langchain-experimental
sqlalchemy==2.0.23
alembic==1.13.1
google-generativeai>=0.5.0
requests>=2.31.0
httpx>=0.27.0
tenacity>=8.2.3
pandas>=2.2.0
tabulate>=0.9.0
matplotlib>=3.8.0
python-dotenv>=1.0.1
pytest>=8.1.1
pytest-asyncio>=0.23.6
respx>=0.21.0
